Added root app option
Prior to this, applications were started in a non-deterministic order,
which could lead to misconfiguration or applications not booting. Now
you can specify a root app that will be started prior to all the other
applications. Due to the way that applications are loaded, specifying
this single app should be enough to boot correctly.